Remove UIntBitMask, switch to android annotations (#31)
To better compatibility with Java

+++ bindings-chicory/src/jvmMain/kotlin/host/module/emscripten/function/SyscallOpenat.kt
@@ -32,7 +32,7 @@ internal class SyscallOpenat(
             memory,
             rawDirFd = args[0].asInt(),
             pathnamePtr = args[1].asWasmAddr(),
-            rawFlags = args[2].asInt().toUInt(),
+            rawFlags = args[2].asInt(),
             rawMode = mode,
         )
         return Value.i32(fdOrErrno.toLong())
